backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
opencv-python
matplotlibclear
python-multipart  # For image upload
